component for viewing episode
