+++ web/src/components/EpisodeViewer.vue
@@ -1,0 +1,38 @@
+<template>
+    <div id="video-player">
+
+    </div>
+</template>
+<script lang="ts" setup>
+import { addError } from '@/utils/error';
+import { onMounted, onUnmounted, onUpdated, ref } from 'vue';
+import { create, Player } from "btv-video-player"
+
+const props = defineProps<{
+    episodeId: string
+}>()
+
+const player = ref(null as Player | null)
+
+const current = ref(null as string | null)
+
+const load = async () => {
+    const episodeId = props.episodeId
+    if (current.value !== episodeId) {
+        current.value = episodeId
+        player.value = await create("video-player", {
+            episodeID: episodeId
+        })
+        if (player.value === null) {
+            addError("No available VOD for this episode")
+        }
+    }
+}
+
+onUpdated(load)
+onMounted(load)
+
+onUnmounted(() => {
+    player.value?.dispose()
+})
+</script>
